Fix by Archie Cobbs: change return type for LongConstantValueTag.getLongValue to long

+++ src/soot/tagkit/LongConstantValueTag.java
@@ -43,7 +43,7 @@ public class LongConstantValueTag extends ConstantValueTag
 	};
     }
 
-    public double getLongValue() {
+    public long getLongValue() {
 	return value;
     }
 }
